many changes
- Material now has lighting, fog (configuration) and "other mode" bits,
added lots of enums for these other modes but haven't confirmed the
layout is as expected.

- Material now has fillColor, primColor, blendColor, envColor. -- color
was renamed to lightColor.

- Material also stores the combiner variables, even though nothings done
with them yet.

- Changed how materials are passed around during f3d processing. When a
branch occurs, the originating material gets updated, otherwise when a
non-branch occurs the originating material stays as the jump is
considered a child.

- Added f3d MOVEWORD, implementation. Currently use it to grab the fog
calculation.

- Fog now uses this calculation grabed against the near and far planes
for the area that is rendering. While the shading looks correct right
now, the color looks a bit off (blue-ish). This could be because of the
clear color though.

- Fast3DScripts, GeoScripts and LevelScripts all conform to a abstract
base for processing, in attempts to error proof what commands are used.
When a CMD_XX command exists in the Runtime sub struct of these classes
but is not exposed in the new GetCommandInfo, the program crashes (on
purpose). A down-side to this is that Runtime is now a class instead of
a struct. Eventually the class instances could be pooled to prevent GC
issues.

- FOV, near and far plane are set from the level's first area on level
change. (Slider does not update, yet)

- Picking state binds, so the colors on bounds and the selection buffer
don't get vertex lit (this also made them not z-test, will correct
later)

- added set level to main form to attempt to consolidate the setting of
level. (its still fragmented inside)

- slight overhaul to GraphicsState/GraphicsInterface as i further
understand whats being used.

- removed the gizmos from drawing

- added FogNear and FogFar to graphics interface. These should be set to
the area's projection near and far plane values. Fog is calculated using
these against the fog multiplier and offset. This somewhat breaks state
change detection (as multiplier and offset could remain the same, but
the fog calculation would be different with different near far values)
-- either way, right now ForceInvalidate is used so its not an issue.