Summary
Camera:
- Rewrote the camera code to be more straight forward
- Calculations are done in world/model coordinates
- Uses Transform under the hood (Translation [Vector3], Rotation [Quat],
Scale [float])
- Completely removed usage of Matrix4.LookAt for calculating the view
matrix. View matrix is simply the inverse of the world/model for the
camera.
- Broke apart the two camera modes (Fly/Orbit) into their own
implementation of ICameraMode.
- ICameraMode is a easy to implement interface for treating updates to
the camera (input/selection/view matrix) without knowing anything about
the Camera object itself.
- Camera makes an effort to not request data from ICameraMode unless
it's forced to or new inputs/selections are made.
- Camera also maintains a status of whether or not changes were made
that require a repaint (less repainting)
- You can now set the camera rotation and position externally, doing so
results in the ICameraMode receiving a zero-input event later on when
neccisary (so long as there was a change in position). This does not (or
should not) break orbit.

Vertex Lighting:
- Added structure Light and containers Lights0, Lights1,... Lights7 and
LightArray. These all contain compatible offsets read by move memory and
or written to by move word. Previously move memory was only pulling the
color value off the first light, now all lights are read. LightArray is
essentially 8 Lights. The reason there are 7 different light container
types and 1 light array type is because the last light is always the
ambient value. Each light container (Lights0..Lights7) has the number of
lights plus one ambient color ( there is a minimum of 1 light so
Lights0/Lights1 are identical except in that Lights0 treats its
properties as if it contains no lights, only ambient).
- Light directions are 100% camera space. When a light direction is to
be set the model view is pushed with identity prior to doing so and
popped after which is probably the most lightweight way to do it.
- Things that aren't vertex lit no longer are lit.

Rendering:
- Geometry not loaded through geo-display (command 21) show again.
Currently a dummy geo node is made.
- The material is never reset thoughout a level script run anymore and
each different script makes sure to pass back whatever it did to it.
This is how its supposed to be done, otherwise lighting information gets
lost.
- Fog color is now correct (Fog density is slightly off/strong)

UI:
- Optimized object / warp list population time.
- Camera now is placed above default start transform when there is one.

Code:
- Added Color4a, where Color4b is ARGb, Color4a is RGBa. They are
otherwise identical
- Added strange fixed point matrix type that I've not really got to
implementing past writing to and reading from it.