Added sorting utility
The Sort method of RenderLayer was temporary and generated so much
garbage that it was noticably making navigation chuggy. Now it is much
smoother.

It uses QuickSort, and I added it in a way that other bi-directional
(doubly linked) lists can utilize it without the hideous array
conversion from before.

Note that the Z sorting is still origin based, which isn't as ideal as
sorting by triangle face. So some things will draw out of order still,
this change is primarily towards making the experience smoother.