Added water box editing
You can now see and modify boxes around. They don't yet draw with their
intended texture/material, for now they just use a clear blue color that
gets less clear when they are selected in the list.

Currently, you cannot select water boxes by clicking on them in the 3d
viewport, they have to be selected from the object/warp list.

Code changes:
-Object3D, Warp, WarpInstant all now reference the area directly,
instead of the level.
-Added IAreaProperty (works similar to ILevelProperty/IRomProperty)
-Instead of calling CMD_04 and CMD_05, I added OpenNode and CloseNode to
GeoScripts.Runtime (which CMD_04 and 05 now invoke)
-CMD_15 now opens and closes a node the same way CMD_13 does
(GeoScripts.Runtime). This means a single node can either have one or
none GeoModel assigned to it and the code should probably be edited to
reflect this (IE, not maintaining a list of models on each node).
-I probably forgot to commit CustomSortedCategoryAttribute.cs 's change
to drop the namespace doing so now.